a few changes I forgot to commit

+++ src/view/LoginView.java
@@ -17,9 +17,9 @@ public class LoginView{
         if(App.accountList.checkAccount(new Account(uname, pwd))){
             System.out.println("Login Successful");
             App.currentUser = new Account(uname,pwd);
-            if(App.currentUser.getIsSeller().intValue()==1){
+            if(App.currentUser.getIsSeller()==1){
                 SellerHomeView.main();
-            }else if(App.currentUser.getIsSeller().intValue()==0){
+            }else if(App.currentUser.getIsSeller()==0){
                 BuyerHomeView.main();
             }else{
                 SellerHomeView.main();
